+++ gorethink.go
@@ -12,6 +12,23 @@ var (
 	Log *logrus.Logger
 )
 
+const (
+	SystemDatabase = "rethinkdb"
+
+	TableConfigSystemTable   = "table_config"
+	ServerConfigSystemTable  = "server_config"
+	DBConfigSystemTable      = "db_config"
+	ClusterConfigSystemTable = "cluster_config"
+	TableStatusSystemTable   = "table_status"
+	ServerStatusSystemTable  = "server_status"
+	CurrentIssuesSystemTable = "current_issues"
+	UsersSystemTable         = "users"
+	PermissionsSystemTable   = "permissions"
+	JobsSystemTable          = "jobs"
+	StatsSystemTable         = "stats"
+	LogsSystemTable          = "logs"
+)
+
 func init() {
 	// Set encoding package
 	encoding.IgnoreType(reflect.TypeOf(Term{}))